tradeLLM/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
yfinance==0.2.35
requests==2.31.0
groq==0.4.1
hyperliquid-python-sdk==0.9.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.1.2
python-jose==3.3.0
pydantic-settings==2.1.0
aiohttp==3.9.1
feedparser==6.0.10